improvement: address requested changes
